added connection filters
