+++ network/p2p/conn_gater.go
@@ -8,18 +8,21 @@ import (
 	"go.uber.org/zap"
 )
 
+// Implementing ConnectionGater interface.
+// https://github.com/libp2p/go-libp2p-core/blob/master/connmgr/gater.go
+
 // InterceptPeerDial is called on an imminent outbound peer dial request, prior
 // to the addresses of that peer being available/resolved. Blocking connections
 // at this stage is typical for blacklisting scenarios
 func (n *p2pNetwork) InterceptPeerDial(id peer.ID) bool {
-	return !n.isPeerBlacklisted(id)
+	return true
 }
 
 // InterceptAddrDial is called on an imminent outbound dial to a peer on a
 // particular address. Blocking connections at this stage is typical for
 // address filtering.
 func (n *p2pNetwork) InterceptAddrDial(id peer.ID, multiaddr ma.Multiaddr) bool {
-	return !n.isAddrBlacklisted(id, multiaddr)
+	return true
 }
 
 // InterceptAccept is called as soon as a transport listener receives an
@@ -45,20 +48,6 @@ func (n *p2pNetwork) InterceptSecured(direction libp2pnetwork.Direction, id peer
 // InterceptUpgraded is called for inbound and outbound connections, after
 // libp2p has finished upgrading the connection entirely to a secure,
 // multiplexed channel.
-//
-// It checks whether we reached peers limit, if we do, accept connection only for relevant peers
 func (n *p2pNetwork) InterceptUpgraded(conn libp2pnetwork.Conn) (bool, control.DisconnectReason) {
 	return true, 0
-}
-
-// isPeerBlacklisted checks if the given peer is blacklisted
-func (n *p2pNetwork) isPeerBlacklisted(id peer.ID) bool {
-	// TODO: implement filtering by peer IDs
-	return false
-}
-
-// isAddrBlacklisted checks if the given address is blacklisted
-func (n *p2pNetwork) isAddrBlacklisted(id peer.ID, multiaddr ma.Multiaddr) bool {
-	// TODO: implement filtering of addresses
-	return false
 }
